+++ ios/App/CapApp-SPM/Package.swift
@@ -16,7 +16,8 @@ let package = Package(
         .package(name: "CapacitorPreferences", path: "../../../node_modules/@capacitor/preferences"),
         .package(name: "CapacitorShare", path: "../../../node_modules/@capacitor/share"),
         .package(name: "FoldExternalDisplay", path: "../../../native-plugins/fold-external-display"),
-        .package(name: "FoldNativeCamera", path: "../../../native-plugins/fold-native-camera")
+        .package(name: "FoldNativeCamera", path: "../../../native-plugins/fold-native-camera"),
+        .package(name: "FoldNdi", path: "../../../native-plugins/fold-ndi")
     ],
     targets: [
         .target(
@@ -28,7 +29,8 @@ let package = Package(
                 .product(name: "CapacitorPreferences", package: "CapacitorPreferences"),
                 .product(name: "CapacitorShare", package: "CapacitorShare"),
                 .product(name: "FoldExternalDisplay", package: "FoldExternalDisplay"),
-                .product(name: "FoldNativeCamera", package: "FoldNativeCamera")
+                .product(name: "FoldNativeCamera", package: "FoldNativeCamera"),
+                .product(name: "FoldNdi", package: "FoldNdi")
             ]
         )
     ]